Don't require newline for failure options
I can't think of any real reason to force the users to type a newline.
Also the prompt makes it look like just typing the single character is
sufficient. Finally, this behavior is inconsistent with watch, where we
do not require new lines.

Also try to make it clear that retry is the default if the user presses
enter.