fix logger class loader leak with workaround fix memory leak when reloading a project
When executing the sbt command reload, some log Appenders stay
registered in the global LogManager, which leads to leaked class
loaders , which leads eventually to an OutOfMemoryException: metaspace.

This fix registers a cleanup method, which clears the global
LogManager's appenders and logs on debug, which Appender it cleared.

The correct fix would imho be a proper lifecycle mechanism that allows
to deregister all appenders where appropriate.

The underlying bug can be reproduced when you start sbt in an empty
directory and observe its metaspace usage.