Validate commands in multiparser
It was reported in https://github.com/sbt/sbt/issues/4808 that compared
to 1.2.8, sbt 1.3.0-RC2 will truncate the command args of an input task
that contains semicolons. To address this, I slightly reworked the
multiparser to only accept multi commands if the each semicolon
delimited part is a valid command. This was achieved by changing
(s.combinedParser & cmdPart) | cmdPart to (s.combinedParser & cmdPart).
This fixes the issue because say the user is running an input task, foo,
with the arguments `bar;baz`, then the multi parser will parse it as:
List("foo bar", "baz"). If "baz" isn't a valid command, it gets rejected.

This isn't a perfect solution, because certain multi commands that
theoretically could work do not (this is illustrated in the last
scripted test), but it should at least resolve #4808.